Add Admin::LinkHelper
Extract link-related helpers to an independent helper.

Keeps AdminHelper from getting out of control. Also adds missing specs
for the variety of records `both_links` accepts.
