+++ spec/helpers/admin_helper_spec.rb
@@ -35,42 +35,6 @@ RSpec.describe AdminHelper do
 
   end
 
-  describe '#both_links' do
-    subject { helper.both_links(record) }
-
-    context 'the record is a known class' do
-      let(:record) { FactoryBot.create(:user) }
-      it { is_expected.to eq(helper.send(:user_both_links, record)) }
-    end
-
-    context 'the record is unsupported' do
-      let(:record) { OpenStruct.new }
-
-      it 'raises an NoMethodError' do
-        expect { subject }.to raise_error(NoMethodError)
-      end
-    end
-  end
-
-  describe '#comment_both_links' do
-
-    let(:comment) { FactoryBot.create(:comment) }
-
-    it 'includes a prominence icon' do
-      expect(comment_both_links(comment)).to include('icon-prominence')
-    end
-
-    it 'includes a link to the comment on the site' do
-      expect(comment_both_links(comment)).to include(comment_path(comment))
-    end
-
-    it 'includes a link to admin edit page for the comment' do
-      expect(comment_both_links(comment)).
-        to include(edit_admin_comment_path(comment))
-    end
-
-  end
-
   describe '#highlight_allow_new_responses_from' do
 
     context 'anybody' do
